splitting up configuration files
